Create Students Marks Sum
You are given an array of integers,marks , denoting the marks scored by students in a class.
•	The alternating elements , marks0, marks2, marks4 ,  and so on denote the marks of boys.
•	Similarly marks1, marks3, marks5, , ,  and so on denote the marks of girls.
The array name, marks , works as a pointer which stores the base address of that array. In other words,  marks contains the address where marks0  is stored in the memory.

The locked code stub reads the elements of  along with . Then, it calls the function marks_summation(marks, gender, number_of_students) to get the sum of alternate elements as explained above and then prints the sum.